Refactor: Timestamped 설정 및 패키지 정리
