+++ src/main/java/com/sparta/todo/comment/controller/CommentController.java
@@ -3,9 +3,9 @@ package com.sparta.todo.comment.controller;
 import com.sparta.todo.comment.dto.CommentRequestDto;
 import com.sparta.todo.comment.dto.CommentResponseDto;
 import com.sparta.todo.comment.sevice.CommentService;
-import com.sparta.todo.exception.ResponseStatusDto;
-import com.sparta.todo.exception.Status;
-import com.sparta.todo.util.UserDetailsImpl;
+import com.sparta.todo.global.exception.ResponseStatusDto;
+import com.sparta.todo.global.exception.Status;
+import com.sparta.todo.global.util.UserDetailsImpl;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.DeleteMapping;
