+++ src/main/java/com/sparta/todo/global/exception/GlobalExceptionHandler.java
@@ -1,4 +1,4 @@
-package com.sparta.todo.exception;
+package com.sparta.todo.global.exception;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
