+++ src/main/java/com/sparta/todo/user/repository/RefreshTokenRepository.java
@@ -1,6 +1,6 @@
 package com.sparta.todo.user.repository;
 
-import com.sparta.todo.util.RefreshToken;
+import com.sparta.todo.global.util.RefreshToken;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Long> {
